done some fixes
