+++ blog_posts/1_first-blog.md
@@ -10,3 +10,24 @@ updatedAt: "2025-04-01T10:00:00Z"
 JavaScript is one of the most popular programming languages in the world. It is used to make web pages interactive and dynamic.
 
 In this article, we’ll cover variables, functions, and basic data types in JavaScript.
+
+What is JavaScript?
+JavaScript is the programming language of the web.
+
+It can update and change both HTML and CSS.
+
+It can calculate, manipulate and validate data.
+
+Why Study JavaScript?
+JavaScript is one of the 3 languages all web developers must learn:
+
+1.  HTML to define the content of web pages
+
+2.  CSS to specify the layout of web pages
+
+3.  JavaScript to program the behavior of web pages
+
+JavaScript Can Change HTML Content
+One of many JavaScript HTML methods is getElementById().
+
+The example below "finds" an HTML element (with id="demo"), and changes the element content (innerHTML) to "Hello
